Refactor Attachments Content Resolver to use DI, rename
